api/src/requirements.txt
rich
slowapi
uvicorn
fastapi
pydantic
aioredis
pre-commit
python-jose
tortoise-orm
python-dotenv
pydantic[email]
python-multipart
tortoise-orm[asyncpg]